+++ lib/client.dart
@@ -174,6 +174,7 @@ class Client {
     }
 
     try {
+      Client client = Client();
       final Map resp = await _methodChannel.invokeMethod('create', {
         'identifier': identifier,
         'seed': seed,
@@ -184,7 +185,6 @@ class Client {
         'ethResolverConfigArray': ethResolverConfigArray,
         'dnsResolverConfigArray': dnsResolverConfigArray,
       });
-      Client client = Client();
       client.address = resp['address'];
       client.publicKey = resp['publicKey'];
       client.seed = resp['seed'];